refactor: Move bulk of ttcalc main back into the direwolf package
Makes cgo wrangling much easier!

Also deduped the agwpe_s structs...
