+++ src/ttcalc_main_test.go
@@ -1,4 +1,4 @@
-package main
+package direwolf
 
 import (
 	"testing"
